+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 #### 2.11.0 (To be released)
 
+#### 2.10.1
+
 * Upgrade to Spring Boot 2.3.0 / Spring Framework 5.2.6
 * Add [springboot-authentication-jar](examples/springboot-authentication-jar) example to demonstrate how to customise `hawtconfig.json` with Spring Boot jar app.
 * Change property `hawtio.proxyWhitelist` to `hawtio.proxyAllowlist`